Worked on docker image.
Moved dockerfile to root directory so it can copy the repository to the docker
image for installation during the build process. Highly edited the config file
to work with the docker image when it is copied over. Edited a python script to
call the correct Python v3 binary.

Worked a lot on the dockerfile itself to get entap to work inside of it.
Configuration now seems to work but actually running EnTAP fails with
TransDecoder errors that could not be understood.