Worked on docker image.
Added flask, flask bootstrap, and flask WTForms to the image so development can
continue using a docker container. The CMD is currently for the development
flask run command but will be changed to production once it is ready. EnTAP
files will be placed in /workspace/entap and it is assumed /workspace will be
the PVC for now.