Worked on flask.
Spent very long time troubleshooting the bootstrap flask module. After a lot of
time realized the following:

1. Bootstrap is impossibly out of date, flask bootstrap uses version 3.3 when
   the latest is 4.5
2. WTForms as a result worked poorly with the out of date bootstrap's wtf quick
   form.

Bootstrap will have to be added from scratch, which is almost completely done.
Creating the form with WTForms proved difficult and resulted in reinventing the
wheel with a custom form class that generated HTML.