Worked on flask.
Formulated basic layout of how the behind the scenes will work; terminology
will be a task manager and tasks. The first task to work on is uploading a
remote database.

Also worked on how to incorporate the upload remote database interface to the
front end with the user. Initially tried a separate page from upload databases
but that does not look good. Will merge it with the already existing upload
databases page.