Worked on flask.
Finished adding back the contaminants page, adding all necessary classes.
Realized the form class was also doing the job of a model class, so the model
part was separated into its own class. As a result for the page to work a view,
form, and model was created. The model class works very well because it can be
responsible for its own initialization.