Worked on flask.
Updated the basic configuration by adding a config model class that handles all
model specific operations, removing them from the form class. Like the contams
model it now handles the initialization of its json file so that was moved out
of the application.

Updated the style output of the contams page to look good with its list of
contaminants.