Worked on flask.
Started adding back the databases page, adding the model and view classes. The
structure and front end for databases are completely overhauled because the old
system of treating it as a single page upload menu page did not work. There is
simply too much stuff going on with the databases to fit it into a single page.
Instead it will have its own top level link and many sub pages provided within
the index link.