Worked on flask.
Added back the upload local system by adding a new page within the main
databases page. Integrated all the code into the databases view class for
uploading a local file.

As a result of trying to get the dropdown menu for database uploads working,
realized bootstrap has updated itself to version 5. This begin a very long
process of going through a lot of existing html code and updating it for version
5. Most of the changes dealt with the navigation bar and form generation.