Worked on flask.
Added back abstract task system but this time it is extremely organized and much
better implemented then the first attempt. The system is separated into tasks
using a common abstract task interface, much like the old system. In the new
system tasks are controlled by the singleton class task controller. How the task
and controller communicate is much more clear and easy to implement than the
first iteration of this system.

Added back the download task, renaming it to remote upland task as that is much
more descriptive. Added back the interface for remote uploading a database file
on the new database flask view class. Created new database upload form used to
begin a remote upload task for a database.

Lastly, how the actual flask application is initialized has changed so there is
no longer a create app method. Instead, the application is simply created
directly in the entapp package init module.