Worked on flask.
Added the gunzip task for gunzipping compressed database files. Added route on
databases view for starting a gunzip task. Modified the databases index page to
list an operations menu for each database listing, adding a remove command and a
gunzip command if it is a gunzipped extension.

Added remove command to databases view for removing a database.