Worked on flask.
Started work on adding the indexing task for databases. Ran into an issue with
how the front end presents itself, where there is a button for each individual
file. This does not work well when it comes to doing many databases at once for
indexing which is how it should work for the user.

After a lot of trial and error figured out a way to use form checkboxes to do a
single operation on multiple files. Currently there is just hacky injections
within the code that verify the approach that works.