Worked on flask.
Finished refactoring the databases view to handle the new form post route for
removing databases, which now allows the user to remove as many as they wish all
at once.

Finished adding the indexing task, which automatically gunzips any database that
is gunzipped, and then indexes each one individually so the user can see more
progress happening. This is important because indexing databases can take a long
time. The indexing command is added as another drop down operation like the
remove operation so it allows for selecting multiple databases to index.