Worked on flask.
Finished up the downloading section with minor tweaks here and there. One tweak
was throwing an error if unzipping a file would overwrite an already existing
file. Another tweak was adding a new config parameter that specifies the number
of threads entap uses for configuration and running.

Added initial inputs model that is very similar to the databases model, which
will handle the modeling of input fasta files.