Worked on flask.
Added log item that represents a single log/debug entry. It only requires the
working directory and date because the filename can be constructed from that
information alone.

Added logs model. This model finds all log/debug entries and adds them to a list
of log items. This list is then chronologically sorted.