Worked on flask.
Modified necessary HTML templates to change configuration page so the form
fields are organized by category, copying the structure presented in the example
EnTAP INI file.

Removed data type option from the configuration form, adding it as a default
line of configuration in the controller.

Modified the inputs model, run form, and run view classes to add support for
frame selection. Now a user can select to enable/disable frame selection and
provide a BAM file from the inputs model it if is enabled. The inputs model now
supports the logical separation of FASTA and BAM file lists used to populate the
run form fields.