Worked on flask.
Finished adding the ability to run EnTAP with frame selection by modifying the
run tack class to accept additional arguments for it. Modified appropriate
method in run view class to initialize the modified run task correctly with its
new arguments, passing the new form field values to it. Added SAM along with BAM
in the inputs model.

Started work on a system to check if EnTAP's initial configuration setup run has
been done or not by checking the existence of the binary database file. Modified
the task controller to check if EnTAP is setup or not, adding that as an
additional check for before request processing. Databases, inputs, and run views
are now all redirected to a new setup page if EnTAP has not been setup yet.
Added setup page to the roots view class.