Updated output to stdout.
Updated the output for running entap in either a run task or setup task to
capture the stdout of the entap application itself instead of getting the tail
of the most current log.

Something is still wrong with buffering, despite best efforts the output of
entap is buffered until the application closes, which makes the output useless.